requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
jinja2==3.1.3
httpx==0.26.0
python-dotenv==1.0.1
